refactor : home화면은 별도 controller로 분리한다
